move scanner.File, WarningReader, etc to zbuf/zio
